requirements.txt
fastapi[standard]==0.135.3
sqlalchemy==2.0.49
pymysql==1.1.2
aiomysql==0.3.2
python-dotenv==1.0.1
